neuralsheild/requirements.txt
pandas
numpy
scikit-learn
joblib
textblob
flask
flask-login
flask-mail
flask-sqlalchemy
waitress
python-dotenv
werkzeug
gunicorn

